fix: address PR review comments for error handler robustness
