+++ app/index.js
@@ -25,30 +25,21 @@ const AutoUpdater = require("./autoUpdater");
 const os = require("node:os");
 const isMac = os.platform() === "darwin";
 
-// Network error patterns that indicate transient connection issues (proxy, tunnel, DNS, etc.)
-// These should not terminate the application — the ConnectionManager will retry automatically.
-const NETWORK_ERROR_PATTERNS = [
-  'ERR_TUNNEL_CONNECTION_FAILED',
-  'ERR_PROXY_CONNECTION_FAILED',
-  'ERR_INTERNET_DISCONNECTED',
-  'ERR_NETWORK_CHANGED',
-  'ERR_CONNECTION_RESET',
-  'ERR_CONNECTION_REFUSED',
-  'ERR_CONNECTION_TIMED_OUT',
-  'ERR_NAME_NOT_RESOLVED',
-];
+const { NETWORK_ERROR_PATTERNS } = require("./config/defaults");
 
 function isNetworkError(message) {
-  return NETWORK_ERROR_PATTERNS.some(pattern => message.includes(pattern));
+  return typeof message === 'string' && NETWORK_ERROR_PATTERNS.some(pattern => message.includes(pattern));
 }
 
 // Top-level error handlers for crash diagnostics
 process.on('uncaughtException', (error) => {
-  if (isNetworkError(error.message)) {
-    console.error('[ERROR] Network-related uncaught exception (not terminating):', { message: error.message });
+  const message = error instanceof Error ? error.message : String(error);
+  const stack = error instanceof Error ? error.stack : undefined;
+  if (isNetworkError(message)) {
+    console.error('[ERROR] Network-related uncaught exception (not terminating):', { message });
     return;
   }
-  console.error('[FATAL] Uncaught exception:', { message: error.message, stack: error.stack });
+  console.error('[FATAL] Uncaught exception:', { message, stack });
   process.exit(1);
 });
